+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New parameters were added to the plugins: limiter, gate and compressor.
 - New command line that allows to hide the window when running PulseEffects in service mode. This will allow tray
   applets to hide our window.
+- Initial support to Equalizer APO presets.
 
 ### Note to packagers